Enable binary compatibility validator for detekt-test and detekt-test-api
