--- detekt-test-utils/src/main/kotlin/io/github/detekt/test/utils/NullOutputStream.kt
@@ -1,9 +1,0 @@
-package io.github.detekt.test.utils
-
-import java.io.OutputStream
-
-internal class NullOutputStream : OutputStream() {
-    override fun write(b: Int) {
-        // no-op
-    }
-}
