+++ detekt-test-utils/src/main/kotlin/io/github/detekt/test/utils/NullPrintStream.kt
@@ -1,5 +1,12 @@
 package io.github.detekt.test.utils
 
+import java.io.OutputStream
 import java.io.PrintStream
 
 class NullPrintStream : PrintStream(NullOutputStream())
+
+internal class NullOutputStream : OutputStream() {
+    override fun write(b: Int) {
+        // no-op
+    }
+}
